requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
plotly>=5.18.0
xlsxwriter>=3.1.0
python-dateutil>=2.8.0
streamlit-authenticator>=0.3.0
pyyaml>=6.0
